Write a C program that prints the size of various types on the computer it is compiled and run on

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	long long int b;
 	float f;
 
-	printf("Size of a char: %lu byte(s)\n", sizeof(c));
-	printf("Size of an int: %lu byte(s)\n", sizeof(i));
-	printf("Size of a long int: %lu byte(s)\n", sizeof(l));
-	printf("Size of long long int: %lu byte(s)\n", sizeof(b));
-	printf("Size of a float: %lu byte(s)\n", sizeof(f));
+	printf("Size of a char: %d byte(s)\n", (int)sizeof(c));
+	printf("Size of an int: %d byte(s)\n", (int)sizeof(i));
+	printf("Size of a long int: %d byte(s)\n", (int)sizeof(l));
+	printf("Size of long long int: %d byte(s)\n", (int)sizeof(b));
+	printf("Size of a float: %d byte(s)\n", (int)sizeof(f));
 	return (0);
 }
